moving resources from index to PostsPage
